backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18

# Database
sqlalchemy>=2.0.30
aiosqlite>=0.20.0

# PDF / OCR
pdfplumber>=0.11.0
Pillow>=10.0.0

# Knowledge Graph
networkx>=3.3

# Report generation
reportlab>=4.2.0

# Data / Utils
pydantic>=2.10.0
python-dotenv>=1.0.0
httpx>=0.28.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
